Extracting components for reusability
