+++ src/components/MainServices.js
@@ -1,30 +1,21 @@
 import PageLink from "components/PageLink";
-import { LargeContainer, FlexRowCenter } from "components/CommonStyles";
+import {
+  FlexRowCenter,
+  LargeContainerFlex,
+  ResponsiveItemsContainer,
+  ResponsiveContainer,
+  ItemTitle,
+  ItemDescription,
+} from "components/CommonStyles";
 import { HighlightRed } from "components/CommonStyles";
-import { Box, Typography } from "@mui/material";
 import { styled } from "@mui/material/styles";
 import { keyframes } from "@mui/styled-engine";
 import { useMediaQuery } from "@mui/material";
-import mergician from "mergician";
 import vaccination from "stories/assets/vaccination.png";
 import toothbrush from "stories/assets/toothbrush.png";
 import bed from "stories/assets/bed.png";
 import ambulance from "stories/assets/ambulance.png";
 import useTriggerOnScroll from "hooks/useTriggerOnScroll";
-
-const Container = styled(LargeContainer)({
-  display: "flex",
-  alignItems: "flex-start",
-  flexWrap: "wrap",
-});
-
-const ServicesListContainer = styled(Box)(({ theme }) => ({
-  width: "100%",
-  padding: theme.spacing(2),
-  paddingRight: "0px",
-  display: "flex",
-  flexWrap: "wrap",
-}));
 
 const ANIMATION_TIME = 0.8;
 
@@ -46,27 +37,14 @@ const bounceIn = keyframes({
   },
 });
 
-const InfoContainer = styled(Box, {
+const InfoContainer = styled(ResponsiveContainer, {
   shouldForwardProp: (prop) => prop !== "trigger" && prop !== "delay",
 })(({ theme, trigger, delay }) => ({
-  minHeight: "300px",
-  background: theme.palette.background.default,
-  display: "flex",
-  flexDirection: "column",
-  justifyContent: "flex-start",
   border: `1px solid ${theme.palette.divider}`,
   boxShadow: theme.shadows[2],
+  background: theme.palette.background.default,
   padding: theme.spacing(2),
   paddingTop: theme.spacing(3),
-  marginRight: theme.spacing(2),
-  marginBottom: theme.spacing(3),
-  maxWidth: "calc(100% - 16px)",
-  [theme.breakpoints.up("sm")]: {
-    maxWidth: "calc(50% - 16px)",
-  },
-  [theme.breakpoints.up("tablet_min_900")]: {
-    maxWidth: "calc(25% - 16px)",
-  },
   opacity: 0,
   animation: trigger
     ? `${ANIMATION_TIME}s ease ${delay}s forwards ${bounceIn}`
@@ -81,25 +59,6 @@ const StyledImg = styled("img")(({ theme }) => ({
   marginLeft: "auto",
   marginRight: "auto",
 }));
-
-const getCommonStyles = ({ theme }) => ({
-  marginBottom: theme.spacing(1.5),
-  color: theme.palette.text.secondary,
-  textAlign: "center",
-});
-
-const Title = styled(Typography)(({ theme }) =>
-  mergician(getCommonStyles({ theme }), {
-    fontWeight: 600,
-    textDecoration: "underline",
-    overflow: "hidden",
-    textOverflow: "ellipsis",
-  })
-);
-
-const Description = styled(Typography)(({ theme }) =>
-  mergician(getCommonStyles({ theme }), { marginBottom: "0px" })
-);
 
 const MainServices = () => {
   const { targetEl: targetEl1, trigger: trigger1 } = useTriggerOnScroll();
@@ -127,19 +86,19 @@ const MainServices = () => {
   };
 
   return (
-    <Container>
-      <ServicesListContainer>
+    <LargeContainerFlex>
+      <ResponsiveItemsContainer>
         <InfoContainer
           ref={targetEl1}
           trigger={trigger1}
           delay={getDelayArr({ max599, between600_899 })[0]}
         >
           <StyledImg src={vaccination} alt='vaccination' />
-          <Title variant='h2'>Vacunación.</Title>
-          <Description>
+          <ItemTitle>Vacunación.</ItemTitle>
+          <ItemDescription>
             Entendemos la importancia de proteger a tu compañero. Porque un
             futuro saludable comienza con una prevención efectiva.
-          </Description>
+          </ItemDescription>
         </InfoContainer>
         <InfoContainer
           ref={targetEl2}
@@ -147,11 +106,11 @@ const MainServices = () => {
           delay={getDelayArr({ max599, between600_899 })[1]}
         >
           <StyledImg src={toothbrush} alt='dental health' />
-          <Title variant='h2'>Salud dental.</Title>
-          <Description>
+          <ItemTitle>Salud dental.</ItemTitle>
+          <ItemDescription>
             Prevenimos problemas dentales y mantenemos una boca saludable para
             una vida larga y feliz.
-          </Description>
+          </ItemDescription>
         </InfoContainer>
         <InfoContainer
           ref={targetEl3}
@@ -159,11 +118,11 @@ const MainServices = () => {
           delay={getDelayArr({ max599, between600_899 })[2]}
         >
           <StyledImg src={bed} alt='admission' />
-          <Title variant='h2'>Internaciones.</Title>
-          <Description>
+          <ItemTitle>Internaciones.</ItemTitle>
+          <ItemDescription>
             Con un equipo altamente capacitado y un ambiente acogedor, tu
             mascota estará en buenas manos durante su recuperación.
-          </Description>
+          </ItemDescription>
         </InfoContainer>
         <InfoContainer
           ref={targetEl4}
@@ -171,16 +130,16 @@ const MainServices = () => {
           delay={getDelayArr({ max599, between600_899 })[3]}
         >
           <StyledImg src={ambulance} alt='emergency services 24 hours' />
-          <Title variant='h2'>Servicio de emergencias 24 hs.</Title>
-          <Description>
+          <ItemTitle>Servicio de emergencias 24 hs.</ItemTitle>
+          <ItemDescription>
             Atendemos tu emergencia las 24/7: estamos aquí para proteger la
             salud de tu mascota.
-          </Description>{" "}
+          </ItemDescription>{" "}
           <HighlightRed sx={{ fontSize: "1.5rem", textAlign: "center" }}>
             Te: 555-555-555
           </HighlightRed>
         </InfoContainer>
-      </ServicesListContainer>
+      </ResponsiveItemsContainer>
       <FlexRowCenter>
         <PageLink
           to='/?path=/story/stories-mainservices--playground'
@@ -192,7 +151,7 @@ const MainServices = () => {
           Solicitar turno
         </PageLink>
       </FlexRowCenter>
-    </Container>
+    </LargeContainerFlex>
   );
 };
